Use int instead of short for the models

+++ Zeebe.Worker/Models/Command/ActivateJobsRequest.cs
@@ -5,7 +5,7 @@ namespace Zeebe.Worker.Models.Command
 {
     public record ActivateJobsRequest(
         [Required] string JobType,
-        [Required] short? MaxJobsToActivate,
+        [Required] int? MaxJobsToActivate,
         string Timeout,
         string WorkerName,
         IList<string> FetchVariables);
