+++ Zeebe.Worker/Models/Command/UpdateJobRetriesRequest.cs
@@ -4,5 +4,5 @@ namespace Zeebe.Worker.Models.Command
 {
     public record UpdateJobRetriesRequest(
         [Required] long? JobKey,
-        short? Retries);
+        int? Retries);
 }
